refactor: Atom / Text Props 타입 + 위치 변경 (#50)
